@@ -13,41 +13,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/fire": "^7.2.1",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
-    "@capacitor/android": "3.4.1",
-    "@capacitor/app": "1.0.7",
-    "@capacitor/core": "3.3.4",
-    "@capacitor/haptics": "1.1.3",
-    "@capacitor/keyboard": "^1.2.0",
-    "@capacitor/status-bar": "1.0.6",
-    "@ionic/angular": "^6.0.0",
-    "firebase": "^9.4.0",
-    "moment": "^2.29.1",
+    "@angular/common": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
+    "@capacitor/android": "4.0.1",
+    "@capacitor/app": "4.0.1",
+    "@capacitor/core": "4.0.1",
+    "@capacitor/haptics": "4.0.1",
+    "@capacitor/keyboard": "4.0.1",
+    "@capacitor/status-bar": "4.0.1",
+    "@ionic/angular": "^6.1.9",
+    "moment": "^2.29.4",
     "ng2-currency-mask": "^13.0.3",
     "ngx-moment": "^6.0.2",
-    "rxfire": "^6.0.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.1",
+    "@angular-devkit/build-angular": "^14.0.0",
     "@angular-eslint/builder": "~13.0.1",
     "@angular-eslint/eslint-plugin": "~13.0.1",
     "@angular-eslint/eslint-plugin-template": "~13.0.1",
     "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.0.1",
-    "@angular/compiler": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0",
-    "@angular/language-service": "~13.0.0",
-    "@capacitor/cli": "3.3.4",
-    "@ionic/angular-toolkit": "^5.0.0",
+    "@angular/cli": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/compiler-cli": "^14.0.0",
+    "@angular/language-service": "^14.0.0",
+    "@capacitor/cli": "4.0.1",
+    "@ionic/angular-toolkit": "^6.0.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -67,7 +64,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.4"
+    "typescript": "~4.7.3"
   },
   "description": "An Ionic project"
 }
